Update Issue_5
Obviously, the p-values in the two cases are the same, but the confidence intervals change. The pairs that differ significantly at α=5% are drug 1-Placebo, drug 2 -Placebo, and drug2-drug1, while the pairs that differ significantly at α=1% are drug1-Placebo and drug 2 -Placebo. We observe that the difference between treatments 2 and 3 is significant at α=5%, whereas it is not significant at α=1%.